app/requirements/production.txt
Django==4.0.5
django-crispy-forms==1.14.0
django-hitcount==1.3.5